% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/accumulate.r
\name{accumulate_long}
\alias{accumulate_long}
\alias{accumulate_mat}
\alias{accumulate_rcpp}
\title{Accumulating \code{long int}s using \strong{Rcpp}}
\usage{
accumulate_long(n)

accumulate_mat(n)
}
\arguments{
\item{n}{number of objects}
}
\description{
Filler description.
}
\examples{
# single values
accumulate_long(5)
n <- 10^ceiling(log(sqrt(.Machine$integer.max), 10))
m <- choose_long(n, 2)
accumulate_long(m)
# matrix entries
test <- accumulate_mat(m)
print(test)
mode(test)
}
