% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/choose.r
\name{choose_int}
\alias{choose_int}
\alias{choose_long}
\alias{choose_rcpp}
\title{Combinatorial combinations in C using \code{int} and \code{long int}}
\usage{
choose_int(n, k)

choose_long(n, k)
}
\arguments{
\item{n}{number of objects}

\item{k}{size of subset}
}
\description{
Filler description.
}
\examples{
# validation
choose(5, 2)
choose_int(5, 2)
choose_long(5, 2)
# behavioral testing
n <- 10^ceiling(log(sqrt(.Machine$integer.max), 10))
choose(n, 2)
choose_int(n, 2)
choose_long(n, 2)
}
